+++ CHANGELOG.md
@@ -8,6 +8,11 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased]
 
+### Added
+
+- Benchmark tests for the gRPC instrumentation. (#296)
+- Integration testing for the gRPC instrumentation. (#297)
+
 ## [0.11.0] - 2020-08-25
 
 ### Added